properly namespace ItemFactory
