+++ lib/munge.rb
@@ -21,7 +21,7 @@ require "munge/readers/filesystem"
 require "munge/core/config"
 require "munge/core/renderer"
 require "munge/core/router"
-require "munge/core/source/item_factory"
+require "munge/core/item_factory"
 require "munge/core/source"
 require "munge/core/transform"
 require "munge/core/transform_scope_factory"
